back to 8

+++ SSTV-Arduino-Scottie1-Library/SSTV-Arduino-Scottie1-Library.cpp
@@ -422,8 +422,8 @@ void scottie1_transmit_file(char* filename){
 
 
 void jpeg_decode(char* filename, char* fileout){
-//  uint8_t *pImg;
-  uint16_t *pImg;
+  uint8_t *pImg;
+//  uint16_t *pImg;
   int x,y,bx,by;
   byte sortBuf[15360]; //320(px)*16(lines)*3(bytes) // Header buffer
   int i,j,k;
